Add initial project configuration and setup
This commit introduces the basic setup and configuration files for the project. It includes Gradle build configuration, Git ignores, project-level settings and directories, and README file. The project is set to use Kotlin version 1.9.21 and JDK 17. A Gradle wrapper and entry scripts for both Unix-based and Windows operating systems are also included.